requirements.txt
aiohttp==3.6.2
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.3.0
Automat==0.6.0
azure-cognitiveservices-vision-face==0.4.0
azure-common==1.1.23
blinker==1.4
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cloud-init==19.2
colorama==0.3.7
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==2.1.4
distro-info===0.18ubuntu0.18.04.1
Flask==1.1.1
httplib2==0.9.2
hyperlink==17.3.1
idna==2.8
idna-ssl==1.1.0
incremental==16.10.1
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.10.3
jsonpatch==1.16
jsonpointer==1.10
jsonschema==2.6.0
keyring==10.6.0
keyrings.alt==3.0
language-selector==0.1
MarkupSafe==1.1.1
msrest==0.6.10
multidict==4.6.1
netifaces==0.10.4
oauthlib==3.1.0
PAM==0.4.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycrypto==2.6.1
pygobject==3.26.1
PyJWT==1.5.3
pyOpenSSL==17.5.0
pyserial==3.4
python-apt==1.6.4
python-debian==0.1.32
pyxdg==0.25
PyYAML==3.12
requests==2.22.0
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
SecretStorage==2.3.1
service-identity==16.0.0
six==1.13.0
slackclient==2.4.0
ssh-import-id==5.7
systemd-python==234
Twisted==17.9.0
typing-extensions==3.7.4.1
ufw==0.36
urllib3==1.25.7
WALinuxAgent==2.2.40
Werkzeug==0.16.0
yarl==1.4.2
zope.interface==4.3.2
